ALLJOYN-1672 update Unity Extention to handle Container Types
This is the initial work done to make it possible to add more
container types to the unity bindings than are currently
supported.  The requirement to handle an array of MsgArgs by
using the MsgArgs class has been removed.  Now, the code only
needs the MsgArg class.  Any code that used MsgArgs has been
marked as obsolete.

Programs that used MsgArgs will still work but will give a compiler warning. Stating that the code is not obsolete.

Unit test code has been added for the new MsgArg class.  The
ability to implicitly convert an array of booleans to a MsgArg
that contains booleans has been removed.  It is still possible
to use an array off booleans but it must be done using the
MsgArg.Set function.  This was done because of an error when
doing implicit conversion.

The SConscript now copies alljoyn_c.dll into the unit test
folder where Visual Studio builds the unit tests.  This
facilitates quicker testing while developing since the unit
tests can be launched from within visual studio.

Updated unit tests to call obsolete versions of the methods
as well as the non-obsolete versions of the methods.

MsgArg.Get method has been added added to the MsgArg Api.

Before adding the MsgArg.Get method to the API the only way to
read a value from the MsgArg was to use the implicit assignment
operators. The problem with using the impicit assignment
operators is that there is no way to git the QStatus.  Also
there was no way to differentiate strings, object paths, and
signatures; since all three are just strings in the code.

Add the ability to set and get a multi dimensional array MsgArg

When setting multi-dimensional arrays the user must use jagged
arrays.

When getting an array value the jagged array is returned as an object with an internal object[].  It is upto the user of the code to access the object array.

Updated the documentation:
Many methods have been added to MsgArg and others were reporting
that they still needed to be documented.  Any remaining code
that had not been document on the alljoyn_unity now has
documentaiton.