White space clean up
The unity code had a very inconsistent usage of white space.  Some lines were
indented using spaces while others were indented using tabs. In a few places
both tabs and spaces were mixed on the same line.

I chose to use have indentation set using tabs. This seems to be the most
widely accepted way for indentation in C# code.  It was also the way the code
was written originally.  If we decide to use spaces its quick to do a global
replace of tabs with spaces.

Fixed indentation of many of the comment blocks.  Removed some unneeded new
lines that were spread throughout the code.

Change-Id: Ib0640e59ee2e68e05719b474e2662c6a2f7dba99