Added method call for GetNumericVersion
Also added unit tests for all of the Version related methods.
	GetExtensionVersion()
	GetVersion()
	GetBuildInfo()
	GetNumericVersion()

The GetEntensionVersion looks like it was added to track the
modifications for the Unity Extension.  It was originally asigned
the value .3 I changed it to 0.0.3 so that it would have a three
layer release value that matches what is used by the rest of
AllJoyn.

Change-Id: If383d68bf24a07eff8e750414df63caa0d8d5675