Added the ability to handle Dictionaries in MsgArg
Currently the code is limited it can currently only handle dictionaries if the
key and value are object type.

The added unit test is only a smoke screen test.  More situations need to be
added to the unit tests to make sure no signatures can break the binding.

KeyStoreListener was updated with the proper delegate signature.  It currently has no
unit tests.

Change-Id: I9d4e94c320024f44e08f4620da7568b7f43831dc