ALLJOYN-1672 Work expand MsgArg to handle all of the container types
Mostly additional test cases to verify that the MsgArg class is working as
expected fixed a few error found as a result of the unit tests.

If a user tries to set a MsgArg with the wrong signature they will get a
AllJoyn.QStatus.BUS_BAD_SIGNATURE error in response.

Added ability to read the MsgArg signature.

Change-Id: I83c17ff74e7a83a14d7bc3b8aa5a44b5bfb35d3b