Additional Updates for the MsgArg
Additional Constructor added to Create a MsgArg array from a native IntPtr and
the number of args in the MsgArg.  This makes it possible to create a C# MsgArg
when a Native C function returns a pointer to the first element in a MsgArg
array.

The GetHashCode function needed to be changed.  It was returning a hash code
based on the string returned from the ToString method.  The problem is that
the ToString method returns an empty string when compiled in Release mode.
This caused the function to return the same hash code for every MsgArg
regardless of its contents.

Updated the ToString unit test to account for the fact that when using release
mode the ToString returns an empty string.

Change-Id: Ic4586364db7912b29e7c0753dfbc6ee478dcfb44