Implemented parts of the ProxyBusObject
Implemented the following methods for the AllJoyn Unity Extension
   - ProxyBusObject.GetInterface
   - ProxyBusObject.ImplementsInterface
   - ProxyBusObject.ParseXml
   - ProxyBusObject.SecureConnection
   - ProxyBusObject.Path [get]
   - ProxyBusObject.ServiceName [get]
   - ProxyBusObject.SessionId [get]
   - ProxyBusObject.IsValid [get]

Unit tests added for the following methods
   - ProxyBusObject Constructor
   - ProxyBusObject.Dispose
   - ProxyBusObject.IntrospectRemoteObject
   - ProxyBusObject.Path
   - ProxyBusObject.ServiceName
   - ProxyBusObject.SessionId
   - ProxyBusObject.ImplementsInterface
   - ProxyBusObject.GetInterface
   - ProxyBusObject.AddInterface
   - ProxyBusObject.ParseXml

Fixed possible error.  If a string is read from the C bindings and that string
returns as a null pointer the code that converts the byte buffer to a string
could return a buffer underflow error when it tried to remove the nul character

Made the constructor for the InterfaceDescription from an IntPtr public so it
can be called from the ProxyBusObject.GetInterface function.

Added the following method to InterfaceDescription
   - InterfaceDescription.Introspect (wanted for testing ProxyBusObject)

Additional minor white space changes.

Change-Id: I29a9323a5ff5fea2771ec397b8121cc2c610714d