ALLJOYN-1838 Added CancelSessionlessMessage method to Unity Extension
Since the CancelSessionlessMessage method is usless without knowing the Message
or the Serial of the Message than should be canceled I had to add BusObject.Signal
methods that take the Message and fill in the serial. The new Signal methods were
only added that accept the MsgArg data type not MsgArgs.  MsgArgs data type is
deprecated.

Fixed documentation error for the RequestNameReply enum.

Unit test to verify that the CancelSessionlessMessage methods are called.  This test
does not verify that sessionless signals are sent it only verifies that
CancelSessionlessMessage is called.

Change-Id: I6a5d76f52676b522f1703a677f0641bcee9ab992