ALLJOYN-2441 Cleaned up SConscript now that we have a smarter Doxygen builder
Since the Doxygen builder now builds up a list of dependencies and targets when
building html output we no longer have to make dummy targets and we no longer
need to manually add the output directory to the list of items that are cleaned.

The files are now built directly into the dist directory.  This is specified in
the OUTPUT_DIRECTORY option in the Doxygen configuration file.

Change-Id: I40e2f3582592e204fb1022c5c1726c20a649e41c