ALLJOYN-2441 Address error building documentation
When the Doxygen builder is run using cygwin it is messing up the directory
paths. For this reason we can not build the documentation directly into the
distribution directory when building the documentation for unity.  We will
have to build the documentation in place and copy them into the distribution
directory.

Change-Id: Ib0fe3a816d471371e41ff98cad676fc3e62af0a1